Update Author-API environment variables with keys location.
This change introduces two new environment variables for the deploy author-api task during the preprod deploy job.

The first environment variable specifies the name of the S3 bucket where the keys file is stored.
The second environment variable overrides the location of the keys file within the container itself.

An additional variable was added with the IAM policy JSON config that allows the container to read the keys from S3.

A small change was also made to correct the publisher url environment variable, which was missing a trailing slash.